Refactor config modules
