+++ crypto_screener/config/__init__.py
@@ -1,3 +1,4 @@
-from crypto_screener.config.config import AppConfig, PpoConfig, app_cfg, ppo_cfg
+from crypto_screener.config.app_config import AppConfig, app_cfg
+from crypto_screener.config.ppo_config import PpoConfig, ppo_cfg
 
 __all__ = ["AppConfig", "PpoConfig", "app_cfg", "ppo_cfg"]
